pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "nonebot_plugin_poker"
version = "1.0.1"
description = "详见https://github.com/MoonofBridge24/nonebot_plugin_poker"
authors = [
    { name = "MoonofBridge24", email = "moonofbridge24@foxmail.com" }
]
dependencies = [
    "nonebot2>=2.3.0",
    "nonebot-adapter-onebot>=2.3.0"
]
readme = "README.md"
requires-python = ">=3.10, <4.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
"Homepage" = "https://github.com/MoonofBridge24/nonebot_plugin_poker"

[tool.setuptools.packages.find]
include = ["nonebot_plugin_poker*"]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
